lambda refactoing
